fix bugs on download modules and output message

+++ Books/ck101.py
@@ -117,9 +117,14 @@ class CK101():
             f.close()
             return True
         def _DownloadChapter(self,url):
-            # open chapter url
-            chRes = urllib.request.urlopen(url,timeout=60)
-            content = chRes.read()
+            try:
+                # open chapter url
+                chRes = urllib.request.urlopen(url,timeout=60)
+                content = chRes.read()
+            except:
+                time.sleep(5)
+                self._DownloadChapter(url)
+                return
             # decode the content
             if (chRes.info().get('Content-Encoding') == 'gzip'):
                 gzipFile = gzip.GzipFile('','rb',9,io.BytesIO(content))
@@ -181,6 +186,7 @@ class CK101():
                         f = False
                         break
                 if(f):
+                    out("\rupdate "+book._name,end="")
                     os.rename(self._path+"\\"+book._bookType+"\\"+book._name+"-"+book._writer+".txt",self._path+"\\"+book._bookType+"\\-"+book._name+"-"+book._writer+".txt")
                     book._name = '-'+book._name
                     self._cursor.execute("update books set name='"+book._name+"' where website='"+book._website+"'")
@@ -201,6 +207,7 @@ class CK101():
         while(errorPage<n):
             b = self.Book("https://www.ck101.org/book/"+str(self._bookNum)+".html")
             if(b._name):
+                out("\r"+b._name,end="")
                 self.books.append(b)
                 flag = bool(self._cursor.execute("select * from books where name='"+b._name+"' and website='"+b._website+"'").fetchone())
                 if(not(flag)):
@@ -209,11 +216,15 @@ class CK101():
                         "('"+b._name+"','"+b._writer+"','"+b._date+"','"+b._chapter+"','"+b._website+"','"+b._bookType+"','false')"
                     )
                     self._cursor.execute(sql)
+                    self._cursor.execute("delete from error where website='"+b._website+"'")
                     self._conn.commit()
                 errorPage = 0
             else:
-                self._cursor.execute("insert into error (website) values ('"+b._website+"')")
-                self._conn.commit()
+                out("\rerror "+str(self._bookNum),end="")
+                f = self._cursor.execute("select * from error where website='"+b._website+"'").fetchone()
+                if(not f):
+                    self._cursor.execute("insert into error (website) values ('"+b._website+"')")
+                    self._conn.commit()
                 errorPage += 1
             self._bookNum += 1
         self.ErrorUpdate()
@@ -222,7 +233,8 @@ class CK101():
         self.books.clear()
         condition = ""
         if(not checkAll): condition = " and type is null or type=''"
-        for row in self._cursor.execute("select * from error where website like '%ck101%'"+condition):
+        rows = self._cursor.execute("select * from error where website like '%ck101%'"+condition).fetchall()
+        for row in rows:
             web = row[0]
             errType = ''
             try:
@@ -240,12 +252,14 @@ class CK101():
             except (urllib.request.socket.timeout): errType = 'timeout'
             self._cursor.execute("update error set type='"+errType+"' where website='"+web+"'")
             if(errType == ''):
-                self._conn.execute("delete from error where website='"+web+"'")
                 b = self.Book(web)
-                sql = (
-                    "insert into books (name,writer,date,chapter,website,type,download) values"
-                    "('"+b._name+"','"+b._writer+"','"+b._date+"','"+b._chapter+"','"+b._website+"','"+b._bookType+"','false')"
-                )
-                self._cursor.execute(sql)
+                if(b._name):
+                    self._conn.execute("delete from error where website='"+web+"'")
+                    out("\rupdate "+b._name,end="")
+                    sql = (
+                        "insert into books (name,writer,date,chapter,website,type,download) values"
+                        "('"+b._name+"','"+b._writer+"','"+b._date+"','"+b._chapter+"','"+b._website+"','"+b._bookType+"','false')"
+                    )
+                    self._cursor.execute(sql)
             self._conn.commit()
         out("error update finish")
